requirements.txt
python-dotenv
fastapi
uvicorn[standard]
python-multipart
supabase
pillow
pymongo
google-generativeai
beautifulsoup4
cloudscraper